+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 <!-- Your new comment below here, and above the most recent release -->
 
+* Add supports GitHub Actions [@duck8823](https://github.com/duck8823)
+
 ## 5.10.3
 
 * Fix issue in `Jenkins` lib with PR detection logic [@serg-kovalev](https://github.com/serg-kovalev)